+++ CHANGELOG.md
@@ -7,9 +7,10 @@ v5.0.0
 ---
 
 * **BREAKING CHANGES**
-    *  The severity level is now configurable and the default severity may have changed:
-        - msbuildFormatter - default was "warning", is now "error"
-        - pmdFormatter - default was priority 1, is not priority 3
+    * The severity level is now configurable, and defaults to severity "error"
+    * The following formatters have change output:
+        - msbuildFormatter - default was "warning"; it is now "error"
+        - pmdFormatter - default was priority 1; it is now "error" (priority 3). If set to "warning", it will output priority 4
 * [enhancement] Enable WARN with new config file format (#629, #345)
     * Valid values for `severity`: "error|warn|warning|none|off"
     * Old style